refactor(CkPmg): split Donut and DebugShapes into per-family files
Review feedback: the module's two feature families lived interleaved in
CkPmg_Utils, CkPmg_Processor, and CkPmg_Fragment_Data. Each is now split into
_Donut and _DebugShapes files (the slim shared Fragment_Data keeps only the
enums both families use — ECk_Pmg_RenderMode, ECk_Pmg_Override); the per-shape
Angular/Basic/Directional/Flat/Icon/Symbol/Text files are unchanged in name.

Get_MeshComponentOuter leaves the fragment files (namespace bleed) and becomes
a typed static per family on its utils class: Donut checks the opt-in tag
directly (standalone shapes), DebugShapes walks the lifetime chain because
composite shapes (icons, dashed lines, text glyph runs) render through child
entities. The Symbol/Text local setup helpers tightened to the typesafe
DebugShape handle — every caller already passes it.

CkCrowd's two PMG consumers repointed at the family headers. Split verified
line-for-line against the deleted files (zero drift); the 6 shape TUs gained
the UMaterialInstanceDynamic include they had been borrowing from the deleted
unity-chunk neighbor.
