+++ Source/CkPmg/Public/CkPmg/CkPmg_Processor_DirectionalShapes.cpp
@@ -8,10 +8,12 @@
 
 #include "CkPmg/CkPmg_Log.h"
 #include "CkPmg/CkPmg_Utils_DebugLines.h"
+#include "CkPmg/CkPmg_Utils_DebugShapes.h"
 #include "CkPmg/CkPmg_Utils_DirectionalShapes.h"
 #include "CkPmg/CkPmg_Utils_FlatShapes.h"
 
 #include <MaterialDomain.h>
+#include <Materials/MaterialInstanceDynamic.h>
 #include <ProceduralMeshComponent.h>
 
 // --------------------------------------------------------------------------------------------------------------------
@@ -118,7 +120,7 @@ namespace ck_pmg_processor_directional_shapes_impl
         { return nullptr; }
 
         auto MeshComponent = UCk_Utils_Object_UE::Request_CreateNewObject<UProceduralMeshComponent>(
-            ck::pmg::Get_MeshComponentOuter(World, InHandle), UProceduralMeshComponent::StaticClass(), nullptr,
+            UCk_Utils_Pmg_DebugShape_UE::Get_MeshComponentOuter(World, InHandle), UProceduralMeshComponent::StaticClass(), nullptr,
             FCk_ObjectPooling_PoolParams{}.Set_RecyclePolicy(ECk_ObjectPooling_RecyclePolicy::DestroyOnRelease), nullptr);
 
         if (ck::Is_NOT_Valid(MeshComponent))
