+++ Source/CkPmg/Public/CkPmg/CkPmg_Processor_FlatShapes.cpp
@@ -8,8 +8,10 @@
 
 #include "CkPmg/CkPmg_Log.h"
 #include "CkPmg/CkPmg_Utils_DebugLines.h"
+#include "CkPmg/CkPmg_Utils_DebugShapes.h"
 
 #include <MaterialDomain.h>
+#include <Materials/MaterialInstanceDynamic.h>
 #include <ProceduralMeshComponent.h>
 
 // --------------------------------------------------------------------------------------------------------------------
@@ -625,7 +627,7 @@ namespace ck_pmg_processor_flat_shapes_impl
         { return nullptr; }
 
         auto MeshComponent = UCk_Utils_Object_UE::Request_CreateNewObject<UProceduralMeshComponent>(
-            ck::pmg::Get_MeshComponentOuter(World, InHandle), UProceduralMeshComponent::StaticClass(), nullptr,
+            UCk_Utils_Pmg_DebugShape_UE::Get_MeshComponentOuter(World, InHandle), UProceduralMeshComponent::StaticClass(), nullptr,
             FCk_ObjectPooling_PoolParams{}.Set_RecyclePolicy(ECk_ObjectPooling_RecyclePolicy::DestroyOnRelease), nullptr);
 
         if (ck::Is_NOT_Valid(MeshComponent))
