+++ Source/CkPmg/Public/CkPmg/CkPmg_Utils_IconShapes.cpp
@@ -1,5 +1,5 @@
 #include "CkPmg_Utils_IconShapes.h"
-#include "CkPmg_Utils.h"
+#include "CkPmg_Utils_DebugShapes.h"
 #include "CkPmg_Fragment.h"
 #include "CkPmg_Fragment_IconShapes.h"
